requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
altair==5.4.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
azure-ai-documentintelligence==1.0.0b4
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.30.2
azure-storage-blob==12.22.0
blinker==1.8.2
Brotli==1.1.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
cssselect2==0.7.0
dataclasses-json==0.6.7
distro==1.9.0
docx2python==3.0.2
docx2txt==0.8
docxpy==0.8.5
exceptiongroup==1.2.2
Faker==30.8.1
fonttools==4.53.1
fpdf==1.7.2
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
groq==0.11.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isodate==0.6.1
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.38
langchain-ollama==0.1.3
langchain-openai==0.1.23
langchain-text-splitters==0.2.2
langsmith==0.1.99
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
msrest==0.7.1
multidict==6.0.5
mypy-extensions==1.0.0
names==0.3.0
narwhals==1.4.1
numpy==1.26.4
oauthlib==3.2.2
ollama==0.3.2
openai==1.40.6
orjson==3.10.7
packaging==24.1
pandas==2.2.2
paragraphs==1.0.1
pillow==10.4.0
protobuf==5.27.3
pyarrow==17.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydyf==0.11.0
Pygments==2.18.0
PyPDF2==3.0.1
pyphen==0.16.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.32
streamlit==1.37.1
tenacity==8.5.0
tiktoken==0.7.0
tinycss2==1.3.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
weasyprint==62.3
webencodings==0.5.1
yarl==1.9.4
zopfli==0.2.3
